Fixed a warning.

+++ tinyuntar/untar.h
@@ -8,7 +8,7 @@
 
 #define IS_BASE256_ENCODED(buffer) (((unsigned char)buffer[0] & 0x80) > 0)
 #define GET_NUM_BLOCKS(filesize) (int)ceil((double)filesize / (double)TAR_BLOCK_SIZE)
-#define GET_LAST_BLOCK_PORTION_SIZE(filesize) (filesize % TAR_BLOCK_SIZE)
+#define GET_LAST_BLOCK_PORTION_SIZE(filesize) (int)(filesize % TAR_BLOCK_SIZE)
 
 #ifdef _MSC_VER
 	#define strtoull _strtoui64
